requirements.txt
azure-ai-inference
praw
assemblyai
moviepy
python-dotenv